### 2021-09-03 #3686 @whedon
[PRE REVIEW]: RandomForestsGLS: An R package for Random Forests for dependent data
**Submitting author:** @ArkajyotiSaha (<a href="http://orcid.org/0000-0001-9671-8196">Arkajyoti Saha</a>)
**Repository:** <a href="https://github.com/ArkajyotiSaha/RandomForestsGLS" target ="_blank">https://github.com/ArkajyotiSaha/RandomForestsGLS</a>
**Version:** v0.1.2
