+++ paper.bib
@@ -1,5 +1,6 @@
 @article{saha2021random,
 	author = {Saha, Arkajyoti and Basu, Sumanta and Datta, Abhirup},
+	doi = {10.1080/01621459.2021.1950003},
 	date-added = {2021-07-15 07:06:58 -0400},
 	date-modified = {2021-07-15 07:06:58 -0400},
 	journal = {Journal of the American Statistical Association},
@@ -17,6 +18,7 @@
 
 @article{breiman1996bagging,
 	author = {Breiman, Leo},
+	doi = {10.1007/bf00058655},
 	journal = {Machine learning},
 	number = {2},
 	pages = {123--140},
@@ -27,6 +29,7 @@
 
 @article{breiman2001random,
 	author = {Breiman, Leo},
+	doi = {10.1023/A:1010933404324},
 	journal = {Machine learning},
 	number = {1},
 	pages = {5--32},
@@ -46,6 +49,7 @@
 
 @article{georganos2019geographical,
 	author = {Georganos, Stefanos and Grippa, Tais and Niang Gadiaga, Assane and Linard, Catherine and Lennert, Moritz and Vanhuysse, Sabine and Mboga, Nicholus and Wolff, El{\'e}onore and Kalogirou, Stamatis},
+	doi = {10.1080/10106049.2019.1595177},
 	journal = {Geocarto International},
 	pages = {1--16},
 	publisher = {Taylor \& Francis},
@@ -55,7 +59,7 @@
 @article{BASAK2019552,
 	abstract = {Predicting returns in the stock market is usually posed as a forecasting problem where prices are predicted. Intrinsic volatility in the stock market across the globe makes the task of prediction challenging. Consequently, forecasting and diffusion modeling undermines a diverse range of problems encountered in predicting trends in the stock market. Minimizing forecasting error would minimize investment risk. In the current work, we pose the problem as a direction-predicting exercise signifying gains and losses. We develop an experimental framework for the classification problem which predicts whether stock prices will increase or decrease with respect to the price prevailing n days earlier. Two algorithms, random forests, and gradient boosted decisio`n trees (using XGBoost) facilitate this connection by using ensembles of decision trees. We test our approach and report the accuracies for a variety of companies as improvement over existing predictions. A novelty of the current work is about the selection of technical indicators and their use as features, with high accuracy for medium to long-run prediction of stock price direction.},
 	author = {Suryoday Basak and Saibal Kar and Snehanshu Saha and Luckyson Khaidem and Sudeepa Roy Dey},
-	doi = {https://doi.org/10.1016/j.najef.2018.06.013},
+	doi = {10.1016/j.najef.2018.06.013},
 	issn = {1062-9408},
 	journal = {The North American Journal of Economics and Finance},
 	keywords = {Stock price movement, Xgboost, Random forests, Machine classification},
