+++ paper.bib
@@ -40,6 +40,7 @@
 
 @article{hengl2018random,
 	author = {Hengl, Tomislav and Nussbaum, Madlene and Wright, Marvin N and Heuvelink, Gerard BM and Gr{\"a}ler, Benedikt},
+	doi = { 10.7717/peerj.5518},
 	journal = {PeerJ},
 	pages = {e5518},
 	publisher = {PeerJ Inc.},
@@ -101,6 +102,7 @@
 
 @article{dagum1998openmp,
 	author = {Dagum, Leonardo and Menon, Ramesh},
+	doi = {10.1109/99.660313},
 	journal = {IEEE computational science and engineering},
 	number = {1},
 	pages = {46--55},
