+++ paper.bib
@@ -113,6 +113,7 @@
 
 @book{breiman1984classification,
 	author = {Breiman, Leo and Friedman, Jerome and Stone, Charles J and Olshen, Richard A},
+	doi = {10.1201/9781315139470},
 	publisher = {CRC press},
 	title = {Classification and regression trees},
 	year = {1984}}
