### 2021-09-05 #3686 @arfon
@ArkajyotiSaha - before this submission is sent out for review, could you please clarify exactly what this statement means, and which code this refers to (from the README)?

> Some code blocks are borrowed from the R packages: spNNGP: Spatial Regression Models for Large Datasets using Nearest Neighbor Gaussian Processes https://CRAN.R-project.org/package=spNNGP and randomForest: Breiman and Cutler's Random Forests for Classification and Regression https://CRAN.R-project.org/package=randomForest

Also, could you say more about what the current paper "Saha, Arkajyoti, Sumanta Basu, and Abhirup Datta. "Random forests for spatially dependent data." Journal of the American Statistical Association (2021): 1-19." discusses?
