### 2021-09-08 #3686 @ArkajyotiSaha
@arfon I have made the changes in README (under Note) to discuss in details the parts of codes which were borrowed from both the packages under consideration. In the overview section of the README, I have also briefly discussed what does our paper "Saha, Arkajyoti, Sumanta Basu, and Abhirup Datta. "Random forests for spatially dependent data." Journal of the American Statistical Association (2021): 1-19." discuss and how does that relate to the package. Please let me know if any additional changes are required.
