### 2021-09-09 #3686 @arfon
Thanks @ArkajyotiSaha. @fabian-s – please take a look at the authors note in the README [here](https://github.com/ArkajyotiSaha/RandomForestsGLS#note) about code reuse. Given the relatively small size of this submission, I believe that this submission may fall foul of our [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion.

@fabian-s – let me know if you'd like help making this assessment.
