### 2021-09-16 #3686 @fabian-s
after reviewing the code, I think I  have to agree with @arfon  's assessment.
