### 2021-09-22 #3686 @ArkajyotiSaha
@fabian-s Attached herewith is a brief summary of the novelty of the code and the method that might be helpful in this review process.

The package implements the RF-GLS algorithm (which is a novel algorithm proposed in "Saha, Arkajyoti, Sumanta Basu, and Abhirup Datta. "Random forests for spatially dependent data." Journal of the American Statistical Association (2021): 1-19.") in R which is a novel Random Forest for correlated data. This method incorporates the correlation structure of the data in regression tree partition and node representative determination by replacing the OLS cost function used in RF with a GLS cost function. Presently there is no package in R that explicitly accounts for a parsimonious modeling of the correlation structure and functional value (mean) estimation for new covariate values in correlated data.

1. Since we use a regression tree, the basic tool and structure is similar to that of Breiman and Cutler's classic Random Forest implemented in `randomForest`. We use just the skeleton of regression trees and associate RF in `randomForest`, as that structure is identical across all regression trees. **The cost function implementation, node representative determination, bootstrapping for each of the regression trees, choice of acceptable splits are handled in a entirely different ways in these two packages and the respective codes in our package are built from scratch with a number of novel implementation optimizations, detailed in the software paper.**

2. Here we model the correlation structure with nearest neighbor Gaussian process (NNGP), which makes the proposed approach scalable. NNGP is implemented in R `spNNGP` in the context of linear models. Here, we use the structure of the code for correlation computation under NNGP, which is identical across all approaches using the NNGP structure. Using NNGP in the cost function used in `findBestSplit` leads to **clever implementation of serial cost function evaluation while searching through possible acceptable gaps, which are built from scratch as these are novel methods with optimized implementation techniques.**

I would be happy to provide with additional details on the novelties of the code and/or the method for the review process. Please let me know.
