### 2021-09-28 #3686 @fabian-s
@cole-brokamp would you be willing to review this implementation of random forests for dependent/spatial data for JOSS?
