### 2021-09-29 #3686 @fabian-s
@pdwaggoner @matloff
would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
