### 2021-09-29 #3780 @whedon
[REVIEW]: RandomForestsGLS: An R package for Random Forests for dependent data
**Submitting author:** @ArkajyotiSaha (<a href="http://orcid.org/0000-0001-9671-8196">Arkajyoti Saha</a>)
**Repository:** https://github.com/ArkajyotiSaha/RandomForestsGLS
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.4.JOSS
