### 2021-11-03 #3780 @mnwright
I think this is a very useful extension of random forests and a promising package. The examples where the methods outperforms standard RF are quite impressing! I have a few general questions, some on the package and some on the paper:

## General
* From what I understand, there are two major differences to standard RF: The bootstrap procedure and the splitting rule. Why not take an existing RF package such as randomForest or ranger and make these changes instead of setting up a new package "borrowing some code from randomForest"?
* Fitting RF-GLS is slower than standard RF. How much is it slower? How does it scale with the number of observations, covariates or other data or model parameters?
* Is a real data example available? That would be of interest for the method itself (not the focus here) but also for the package to see how it scales and for which real data purpose it can be used.

## Package
* The C++ code is not documented/commented well and hard to understand.
* The DESCRIPTIONS still contains a link to arxiv, not the published paper.
* The README is missing a link to the paper.
* Typo in README: criterion.
* Tests just run examples and check output types/sizes. That could be improved with more tests and tests that check for correct output.
* Is any kind of continuous integration used? I think it is useful to at least run the tests with each commit/PR.
* Maybe too late to change that, but I think the package name is not a great choice. For example, at first try, I typed "randomForestGLS", then capitalized to "RandomForestGLS" and finally corrected to "RandomForestsGLS". It's also quite long and you have to remember the capitalization.

## Paper:
* The JASA paper is called "Random Forests for Spatially Dependent Data", the software has the same name without "Spatially". Are additional types of dependencies covered by the software, not described in the original JASA paper? If yes, please detail in the JOSS paper.
* line 8: Should be "in these models"
* lines 14-15: "hence is not optimal in mixed-model approach". I don't understand this. Wouldn't RF be used as an alternative to the mixed model and not IN the mixed model approach?
* lines 18-19: Avoid linebreak in package name
* line 63: for or model correlation?
* line 176: optimizing a cost function (missing a)
* line 211: of the RF-GLS method (missing the)
* line 216: "Efficient implementation thorough" should be through?
* line 217: Maybe remove "clever"?
* References: Datta et al. is in title case, others in sentence case.
* In general, many spelling errors, missing articles, etc.
