+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: RandomForestsGLS
 Type: Package
 Title: Random Forests for Dependent Data
-Version: 0.1.2
+Version: 0.1.3
 Authors@R: c(person("Arkajyoti", "Saha", role=c("aut", "cre"), email="arkajyotisaha93@gmail.com"), person("Sumanta", "Basu", role="aut", email="sumbose@cornell.edu"), person("Abhirup", "Datta", role="aut", email="abhidatta@jhu.edu"))
 Maintainer: Arkajyoti Saha <arkajyotisaha93@gmail.com>
 Author: Arkajyoti Saha [aut, cre],
@@ -10,7 +10,7 @@ Author: Arkajyoti Saha [aut, cre],
 Depends: R (>= 3.3.0)
 Imports: BRISC, parallel, stats, matrixStats, randomForest, pbapply
 Suggests: knitr, rmarkdown, ggplot2, testthat (>= 2.1.0)
-Description: Fits non-linear regression models on dependant data with Generalised Least Square (GLS) based Random Forest (RF-GLS) detailed in Saha, Basu and Datta (2020) <arXiv:2007.15421>.
+Description: Fits non-linear regression models on dependant data with Generalised Least Square (GLS) based Random Forest (RF-GLS) detailed in Saha, Basu and Datta (2021) <doi:10.1080/01621459.2021.1950003>.
 License: GPL (>= 2)
 URL: https://github.com/ArkajyotiSaha/RandomForestsGLS
 BugReports: https://github.com/ArkajyotiSaha/RandomForestsGLS/issues
