### 2021-12-05 bf4f374
Use Github Actions for CI
