### 2021-12-05 60a85cb
Added tinytex to build vignette
