+++ .github/workflows/R-CMD-check.yaml
@@ -32,6 +32,8 @@ jobs:
       - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-pandoc@v1
+      
+      - uses: r-lib/actions/setup-tinytex@master
 
       - uses: r-lib/actions/setup-r@v1
         with:
