+++ paper.bib
@@ -79,7 +79,7 @@
 	journal = {Journal of the American Statistical Association},
 	number = {514},
 	pages = {800--812},
-	title = {{Hierarchical Nearest-Neighbor Gaussian Process Models for Large Geostatistical Datasets}},
+	title = {{Hierarchical nearest-neighbor gaussian process models for large geostatistical datasets}},
 	url = {http://dx.doi.org/10.1080/01621459.2015.1044091},
 	volume = {111},
 	year = {2016},
