+++ src/RFGLS.cpp
@@ -51,6 +51,7 @@ void updateBF_org(double *B, double *F, double *c, double *C, double *D, double 
 #ifdef _OPENMP
     threadID = omp_get_thread_num();
 #endif
+    //compute the covariance matrix of the neighbors
     if(i > 0){
       for(k = 0; k < nnIndxLU[n+i]; k++){
         c[nnIndxLU[i]+k] = spCor(d[nnIndxLU[i]+k], theta[1], nu, covModel, &bk[threadID*nb]);
@@ -61,6 +62,7 @@ void updateBF_org(double *B, double *F, double *c, double *C, double *D, double 
           }
         }
       }
+      //compute B & F, folloing BRISC
       F77_NAME(dpotrf)(&lower, &nnIndxLU[n+i], &C[CIndx[i]], &nnIndxLU[n+i], &info); if(info != 0){error("c++ error: dpotrf failed\n");}
       F77_NAME(dpotri)(&lower, &nnIndxLU[n+i], &C[CIndx[i]], &nnIndxLU[n+i], &info); if(info != 0){error("c++ error: dpotri failed\n");}
       F77_NAME(dsymv)(&lower, &nnIndxLU[n+i], &one, &C[CIndx[i]], &nnIndxLU[n+i], &c[nnIndxLU[i]], &inc, &zero, &B[nnIndxLU[i]], &inc);
@@ -74,9 +76,9 @@ void updateBF_org(double *B, double *F, double *c, double *C, double *D, double 
 }
 
 extern "C" {
-
+  
   double pinv_dgelsy_rss_cpp(double *A, double *b, int nrowA, int ncolA){
-
+  //computes the optimized minimum-norm of the least square problem, i.e. minimize||Ax - b|| with QR factorization
 
     int inc = 1;
     int nlengthb=nrowA;
@@ -139,7 +141,7 @@ extern "C" {
   }
 
   void pinv_dgelsy_beta_cpp(double *A, double *b, int nrowA, int ncolA, double *beta){
-
+  //computes the solution of minimum-norm least square problem, i.e. arg min||Ax - b|| with QR factorization
 
     int inc = 1;
     int nlengthb=nrowA;
@@ -186,7 +188,7 @@ extern "C" {
   }
 
   double pinv_dgelsd_rss_cpp(double *A, double *b, int nrowA, int ncolA){
-
+  //computes the optimized minimum-norm of the least square problem, i.e. minimize||Ax - b|| with SVD
     int inc = 1;
     int nlengthb=nrowA;
     int nlengthA=nrowA * ncolA;
@@ -231,7 +233,7 @@ extern "C" {
   }
 
   void pinv_dgelsd_beta_cpp(double *A, double *b, int nrowA, int ncolA, double *beta){
-
+  //computes the solution of minimum-norm least square problem, i.e. arg min||Ax - b|| with SVD
     int inc = 1;
     int nlengthb=nrowA;
     int nlengthA=nrowA * ncolA;
@@ -247,7 +249,7 @@ extern "C" {
     /* Local arrays */
     double *s = NULL;
     s = new double [nrowA];
-    //double s[nrowA];
+    
     int lwork = -1;
     int liwork;
     int* iwork = NULL;
@@ -369,11 +371,9 @@ extern "C" {
       }
     }
 
-    //SEXP j_r; PROTECT(j_r = allocVector(INTSXP, 1)); nProtect++; INTEGER(j_r)[0] = j;
 
-    //SEXP D_r; PROTECT(D_r = allocVector(REALSXP, j)); nProtect++; double *D = REAL(D_r);
     double *D = (double *) R_alloc(j, sizeof(double));
-
+    //compute the pairwise distances
     for(i = 0; i < n; i++){
[…]
