### 2021-12-20 #3780 @ArkajyotiSaha
@mnwright Extremely sorry! Thanks so much for pointing this out. I have somehow missed updating the version of RFGLS.cpp in the GitHub with the commented version. I have now updated the version. The new comments can be found in the updated version of RFGLS.cpp. For your convenience, I am also adding a link to the latest edit history, which highlights the comments at https://github.com/ArkajyotiSaha/RandomForestsGLS/commit/8e1b11236274844ec92be371d10d8d08dfd33844 . Please let me know if this works. Thanks again!
