### 2021-12-20 #3780 @fabian-s
@ArkajyotiSaha
It seems both reviewers are now satisified with the software and the paper, thank you for taking on these revisions.

However, JOSS submissions are typically 500-1000 words, while your article currently has >3000 words. Please shorten as much as you can.
Especially l. 180 - 260 contain much more technical and mathematical detail than we expect for a JOSS paper, please remove anything that is not a high-level summary of what the software offers from a user perspective as well as anything that is also contained in more detail in your package vignettes or your methodological papers.
