### 2021-12-21 #3780 @ArkajyotiSaha
Thank you so much @fabian-s . I am working on the article to shorten it such that it doesn't take anything away from a high-level summary, but does away with the intricate technical and mathematical details by referring to existing associated literature in package vignettes/methodological paper/github repo description of the package.  Will update you once I am done with the update, thanks!
