+++ paper.md
@@ -20,7 +20,7 @@ affiliations:
    index: 2
  - name: Department of Biostatistics, Johns Hopkins Bloomberg School of Public Health
    index: 3
-date: 5 December 2021
+date: 14 January 2022
 bibliography: paper.bib
 ---
 
@@ -113,13 +113,6 @@ The spatial predictions at new locations `Coordstest`, can be obtained through
 RFGLS_predict_spatial <- RFGLS_predict_spatial(est, Xtest, Coordstest)
 ```
 
-### Unknown model parameters
-More often than not, the true covariance parameters of the data are not known apriori. Given a choice from a prespecified list of covariance functions (software presently allows for exponential, spherical, Matérn and Gaussian covariances), the software accommodates parameter estimation with the argument `param_estimate = TRUE`. For this, first we fit a classical RF ignoring the correlation structure. Next, we fit a zero mean Gaussian process with the covariance structure of choice on the residual to obtain the estimate of the model parameters through maximum likelihood estimation. This initial RF fitting and estimation of the spatial parameters from the RF residuals is justified by theoretical results in @saha2021random proving that even under dependence, naive RF is consistent (although empirically suboptimal). The estimated model parameters are then used to fit RF-GLS. As demonstrated in @saha2021random, this leads to significant improvement over classical RF both in terms of estimation and prediction accuracy. This procedure is again analogous to the linear model setup, where fitting a feasible GLS model  often involves pre-estimating the covariance parameters from residuals from an OLS fit.
-
-It has also been demonstrated (Section 4.4 in @saha2021random) that even under misspecification in covariance structure (i.e. when the true spatial effects are generated from a different covariance structure than the specified covariance for estimation or arbitrary smooth function), using RF-GLS with exponential covariance (the default choice of covariance function in the software) leads to noticeable improvement over simply using classical RF. 
-
-One additional consideration for model parameter estimation involves the computational complexity of the maximum likelihood estimation of Gaussian process. In its original form, this involves $O(n^3)$ computation and $O(n^2)$ storage space, which may be prohibitive for large $n$. This problem is thoroughly studied in spatial statistics literature (we refer the readers to @nngp for a detailed review). In the present scenario, we perform a fast, linear-time estimation of the model parameters with BRISC [@brisc], which is directly implemented in the present software through the R package BRISC (https://CRAN.R-project.org/package=BRISC).
-
 ## Autoregressive (AR) Time Series Data
 ### Model
 
@@ -140,40 +133,13 @@ With response `y` and the covariates `x`, we can fit RF-GLS for temporal data as
 est <- RFGLS_estimate_timeseries(y, x)
 ```
 
-
 ### Prediction of covariate effects
 Prediction of covariate effects in AR process models are similar to that of spatial data and are performed with `RFGLS_predict` as in the case of spatial data.
 
 
-### Unknown model parameters
-For time-series data, most often the model parameters for the autoregressive process (order of autoregression and the autocorrelation coefficients) are unknown. These are estimated using a strategy similar to the spatial case. With the option `param_estimate = TRUE`, the software estimates the model coefficients by fitting a zero mean autoregressive process of user defined order on the residuals from a naive RF fit. The autoregressive parameter fitting is done using `arima` from base R package `stats`.
-
-
-
-### Parallelization
-
[…]
