### 2022-01-14 #3780 @ArkajyotiSaha
@fabian-s I have updated the text. We have moved most of 180 - 260 in package README. Since the analysis of time series data is an important part of the package, and is not highlighted in the methodological paper, we wanted to keep it in this article. In order to further reduce the length of the article, we have alsomoved the discussion on unknown parameter estimation to README section. Please let me know if this works, thanks!
