### 2022-01-17 #3780 @fabian-s
Thanks @ArkajyotiSaha , this is much better already but we're still at ~2000 words.

I think by consistently cutting out all verbosity and rhetorical flourishes (e.g. by following some of the tips here: https://redwoodink.com/resources/10-tricks-to-reduce-your-word-count-in-academic-writing) you could easily reduce that by another 20-40% without losing any relevant conten -- could I ask you to try to do so?
I'm sorry for all this back and forth, I should have raised this problem much sooner during the review.

As an example for the kind of editing you could do, consider my attempt at making your Discussion section more concise:

Original (240 words):
> In this package, we have developed an efficient, parallel implementation of the RF-GLS method proposed in @saha2021random. This accounts for the correlation in the data by incorporating it in the node splitting criteria and the node representative update rule. The package accounts for spatial correlation, modeled with Matérn GP and serial autocorrelation. More often than not, the model parameters are unknown to the users, hence the package has inbuilt parameter estimation corresponding to both the covariance structures. Efficient implementation through C/C++ takes advantage of the NNGP approximation and scalable node splitting update rules which reduces the execution time. For details regarding package features, unknown parameter estimation and parallelization, we refer the readers to the package [README](https://github.com/ArkajyotiSaha/RandomForestsGLS#readme). Present implementation of RF-GLS is slower than that of RF due to the additional computational complexity associated with the split criteria evaluation. In RF-GLS, evaluation of cost function, corresponding to a potential split has $O(t^3)$ computational complexity, where $t$ is the number of leaf nodes at the present split. For very deep trees, this would lead to significant added computational burden over RF, where this step has $O(1)$ complexity. As a future extension of the package, implementation of covariate specific parallel optimization corresponding to each node splitting and improving overall computational complexity of the algorithm can be of independent research interest. We also plan to implement additional forms of time series dependency and perform thorough empirical validation, prior to making it available in the software.

Concise (135 words):
> This package provides an efficient, parallel implementation of the RF-GLS method proposed in @saha2021random which accounts for correlated data with modified node splitting criteria and node representative update rule. The package accounts for spatial correlation via Matérn GP or serial autocorrelation. It provides parameter estimation for both covariance structures. Efficient implementation through C/C++ takes advantage of the NNGP approximation and scalable node splitting update rules which reduces the execution time. More details on package features, parameter estimation and parallelization can be found in the package [README](https://github.com/ArkajyotiSaha/RandomForestsGLS#readme). Since the computational complexity of evaluating potential splits is cubic in the number of leaf nodes for RF-GLS, but constant for standard RF, improving the computational complexity of the algorithm is of independent research interest. We also plan to implement and validate additional forms of time series dependency.
