+++ paper.md
@@ -20,39 +20,28 @@ affiliations:
    index: 2
  - name: Department of Biostatistics, Johns Hopkins Bloomberg School of Public Health
    index: 3
-date: 14 January 2022
+date: 13 February 2022
 bibliography: paper.bib
 ---
 
 # Summary
-With the modern advancements in geographical information systems, remote sensing technologies, low-cost sensors, we are increasingly encountering massive datasets indexed by geo-locations and time-stamps. Modeling such high throughput spatio-temporal data needs to carefully account for spatial/serial dependence in the data, i.e., model the structures (patterns) of the data along space or time. A general model for describing dependent observations $(y_1, y_2, \cdots, y_n)$ and covariates $(\mathbf x_1,\ldots,\mathbf x_n)$ can be formulated as $y_i = f(\mathbf x_i) + \epsilon_i$, where $f(\mathbf x_i)$ denotes the mean component and $\epsilon_i$ is the residual error, which accounts for the dependency in the data, beyond what is explained by the covariates. In this article, we work with second order dependency, i.e. we assume that the dependence is captured through the covariance function of the correlated stochastic process $\epsilon_i$. Since the data are indexed by locations (spatial data) or time-points (temporal data), the correlation is a function of the "spatial distance" or "time-lag" between two observations.
+With the modern advancements in geographical information systems, remote sensing technologies, low-cost sensors, we are increasingly encountering datasets, where we need to account for spatial/serial dependence. Dependent observations $(y_1, y_2, \cdots, y_n)$ with covariates $(\mathbf x_1,\ldots,\mathbf x_n)$ can be modeled non-parametrically as $y_i = m(\mathbf x_i) + \epsilon_i$, where $f(\mathbf x_i)$ is mean component and $\epsilon_i$ accounts for the dependency in data. We assume, dependence is captured through covariance function of correlated stochastic process $\epsilon_i$ (second order dependence). The correlation is typically a function of "spatial distance" or "time-lag" between two observations. 
 
-The primary purpose of the package `RandomForestsGLS` is to fit nonlinear regression for spatial and temporal data. This package performs estimation through a novel rendition of Random Forests (RF); namely, RF-GLS, which makes use of the dependence structure in the data. With increasing computing capacity of personal computers, non-linear Machine Learning (ML) methods have become increasingly commonplace for regression and classification tasks due to their ability of capture complex interactions among the variables, that cannot be modeled by linear regression. However, many modern ML software lack the capability to efficiently account for the dependence structure in the data which leads to sub-optimal estimation. On the other hand, specialized software for spatial/temporal data are capable of properly modeling data correlation, but usually assume the relationships between the response and the covariates.  `RandomForestsGLS` brings the best of both worlds together by bridging the gap between these two approaches by explicitly modeling the spatial/serial data correlation in the RF fitting procedure to substantially improve estimation of the regression function. In ML workflow, `RandomForestsGLS` can substitute existing ML methods in model training to take care of the dependence structure. Similarly, for traditional spatial modeling  frameworks, `RandomForestsGLS` can be used instead of linear model based methods to account for nonlinearity. Additionally, `RandomForestsGLS` seamlessly leverages kriging to perform predictions at new locations for geo-spatial data, a primary objective in many spatial analysis.
[…]
