+++ paper.md
@@ -25,23 +25,23 @@ bibliography: paper.bib
 ---
 
 # Summary
-With the modern advancements in geographical information systems, remote sensing technologies, low-cost sensors, we are increasingly encountering datasets, where we need to account for spatial/serial dependence. Dependent observations $(y_1, y_2, \cdots, y_n)$ with covariates $(\mathbf x_1,\ldots,\mathbf x_n)$ can be modeled non-parametrically as $y_i = m(\mathbf x_i) + \epsilon_i$, where $f(\mathbf x_i)$ is mean component and $\epsilon_i$ accounts for the dependency in data. We assume, dependence is captured through covariance function of correlated stochastic process $\epsilon_i$ (second order dependence). The correlation is typically a function of "spatial distance" or "time-lag" between two observations. 
+With the modern advances in geographical information systems, remote sensing technologies, and low-cost sensors, we are increasingly encountering datasets where we need to account for spatial or serial dependence. Dependent observations $(y_1, y_2, \cdots, y_n)$ with covariates $(\mathbf x_1,\ldots,\mathbf x_n)$ can be modeled non-parametrically as $y_i = m(\mathbf x_i) + \epsilon_i$, where $f(\mathbf x_i)$ is mean component and $\epsilon_i$ accounts for the dependency in data. We assume that dependence is captured through a covariance function of the correlated stochastic process $\epsilon_i$ (second order dependence). The correlation is typically a function of "spatial distance" or "time-lag" between two observations. 
 
-Unlike linear regressions, non-linear Machine Learning (ML) methods for estimating the regression function $m$ can capture complex interactions among the variables. However, they often fail to account for the dependence structure, resulting in sub-optimal estimation. On the other hand, specialized software for spatial/temporal data properly models data correlation, but lacks flexibility in modeling mean function $m$ by restricting to linear models. `RandomForestsGLS` bridges the gap through a novel rendition of Random Forests (RF); namely, RF-GLS, by explicitly modeling the spatial/serial data correlation in the RF fitting procedure to substantially improve estimation of the mean function. Additionally, `RandomForestsGLS` leverages kriging to perform predictions at new locations for geo-spatial data.
+Unlike linear regression, non-linear Machine Learning (ML) methods for estimating the regression function $m$ can capture complex interactions among the variables. However, they often fail to account for the dependence structure, resulting in sub-optimal estimation. On the other hand, specialized software for spatial/temporal data properly models data correlation, but lacks flexibility in modeling the mean function $m$ by restricting to linear models. `RandomForestsGLS` bridges the gap through a novel rendition of Random Forests (RF) -- namely, RF-GLS -- by explicitly modeling the spatial/serial data correlation in the RF fitting procedure to substantially improve estimation of the mean function. Additionally, `RandomForestsGLS` leverages kriging to perform predictions at new locations for geo-spatial data.
 
 # Statement of need
[…]
