### 2022-02-14 #3780 @fabian-s
Thanks -- almost done here, finally!
Please take a look at the typos / grammatical errors I fixed at https://github.com/fabian-s/RandomForestsGLS/pull/1, incorporate as necessary/sensible, then please do run another spell check and proof-read closely and ping me again.
