+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: RandomForestsGLS
 Type: Package
 Title: Random Forests for Dependent Data
-Version: 0.1.3
+Version: 0.1.4
 Authors@R: c(person("Arkajyoti", "Saha", role=c("aut", "cre"), email="arkajyotisaha93@gmail.com"), person("Sumanta", "Basu", role="aut", email="sumbose@cornell.edu"), person("Abhirup", "Datta", role="aut", email="abhidatta@jhu.edu"))
 Maintainer: Arkajyoti Saha <arkajyotisaha93@gmail.com>
 Author: Arkajyoti Saha [aut, cre],
