### 2022-02-20 #3780 @ArkajyotiSaha
I made minor updates in the package vignette and updated the package version.

1. The version tag of the tagged release in GitHub is v0.1.4 .
2. The archived software in Zenodo has the DOI 10.5281/zenodo.6192456 .

Please let me know if any additional information is needed.
