### 2022-02-21 #3780 @fabian-s
@ArkajyotiSaha thanks, almost done,

please make sure the title of your Zenodo deposit matches the title of your JOSS paper -- it's

"ArkajyotiSaha/RandomForestsGLS: v0.1.4"

right now, it should be

"RandomForestsGLS: An R package for Random Forests for dependent data"
