### 2022-02-21 #3780 @fabian-s
oh no.....

I'm sorry, @ArkajyotiSaha, but you still need to fix this typo

"the binary membersip matrices" --> members**h**ip

please do run another spell check as well to make sure we don't miss anything,
then make another tagged release and modify your zenodo deposit accordingly and ping me again here once that's done.

sorry again for the back and forth.
