### 2022-02-24 #3780 @ArkajyotiSaha
@fabian-s I have corrected the typo, given the paper multiple through reads and corrected some minor grammatical issues.

1. The version tag of the tagged release in GitHub is v0.1.4.JOSS  .
2. The archived software in Zenodo has the DOI 10.5281/zenodo.6252368 .

Please let me know if any additional information is needed.
