### 2022-02-24 #3780 @fabian-s
@ArkajyotiSaha

1. your Zenodo deposit is wrong again:

![image](https://user-images.githubusercontent.com/998541/155527112-62b23e5e-ff9a-487b-b9f6-ff0f7955c886.png)

looks like the meta data got messed up again....

please DO make sure your Zenodo deposit has

-  the same title as the JOSS paper
-  the same authors  as the JOSS paper

2.
just ignore the bot's complaint about the invalid DOI, the reference you supply:

> Breiman, L. (2001). Random forests.Machine Learning,45(1), 5–32. https://doi.org/10.1251023/A:1010933404324

works, the bot seems to ignore the colon after A and that breaks the DOI.
