### 2022-02-24 #3780 @fabian-s
Attention, EiC --
the bot's complaint about the invalid DOI is a false poitive: the supplied reference:

> Breiman, L. (2001). Random forests.Machine Learning,45(1), 5–32. https://doi.org/10.1251023/A:1010933404324

works, but the bot seems to ignore the colon after A and that breaks the DOI.
