### 2022-02-24 #3780 @arfon
@fabian-s - looks like @editorialbot is failing to compile the paper due to some issues with the maths in the doc. / cc @tarleb  for visibility here.
