### 2022-02-24 #3780 @arfon
@xuanxu looks like the DOI `10.7717/peerj.5518` is causing the XML validation to fail here (although it's clearly a valid DOI). Perhaps we shouldn't fail the workflow here?

@fabian-s @ArkajyotiSaha, apologies for the noise here. We switched over to the new bot infrastructure yesterday and you're helping us find a few bugs 😅
