### 2022-02-24 #3780 @tarleb
The problem was that the DOI has a leading space in the bib file. Our pipeline should be able to auto-fix minor issues like that. I've pushed a change that should allow validation to succeed now.
