### 2022-03-07 #3780 @ArkajyotiSaha
Hi @fabian-s I just noticed that the link to the software archive at https://joss.theoj.org/papers/10.21105/joss.03780 is not working. This directs to 10.21105/zenodo.6257157 instead of the original zenodo archive at 10.5281/zenodo.6257157 . I was wondering if there is a way to fix this. Thanks!
